forge fmt

+++ src/adapters/V2Adapter.sol
@@ -10,40 +10,30 @@ contract V2Adapter {
         router = IUniswapV2Router(_router);
     }
 
-    function quote(
-        address tokenIn,
-        address tokenOut,
-        uint amountIn
-    ) external view returns (uint amountOut) {
+    function quote(address tokenIn, address tokenOut, uint256 amountIn) external view returns (uint256 amountOut) {
         address[] memory path = new address[](2);
         address;
         path[0] = tokenIn;
         path[1] = tokenOut;
 
-        try router.getAmountsOut(amountIn, path) returns (uint[] memory amounts) {
+        try router.getAmountsOut(amountIn, path) returns (uint256[] memory amounts) {
             amountOut = amounts[1];
         } catch {
             amountOut = 0;
         }
     }
 
-    function swap(
-        address tokenIn,
-        address tokenOut,
-        uint amountIn,
-        uint minOut
-    ) external returns (uint amountOut) {
+    function swap(address tokenIn, address tokenOut, uint256 amountIn, uint256 minOut)
+        external
+        returns (uint256 amountOut)
+    {
         address[] memory path = new address[](2);
         address;
         path[0] = tokenIn;
         path[1] = tokenOut;
 
-        uint[] memory amounts = router.swapExactTokensForTokens(
-            amountIn,
-            minOut,
-            path,
-            address(this),
-            block.timestamp
+        uint256[] memory amounts = router.swapExactTokensForTokens(
+            amountIn, minOut, path, address(this), block.timestamp
         );
 
         return amounts[1];
